Add isNull function.
Null elements are those that are just default constructed. Any subsequent
non-const operation (putId, putTarget, etc) makes the element no
null. Knowing when an element is null is useful for queries. E.g.: if we ask
a document for an element of particular id, and the element is not there,
the doc will return a null element (join, link, fs or word)